requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==10.0.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
ctranslate2==3.24.0
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
decorator==5.2.1
distro==1.9.0
dtw-python==1.4.4
edge-tts==7.0.2
elevenlabs==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.3.0
fastapi==0.115.13
ffmpy==0.6.0
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gradio_client==0.6.1
groq==0.28.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.43.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.5.0
murmurhash==1.0.13
narwhals==1.44.0
networkx==3.4.2
numba==0.60.0
numpy==1.22.0
oauthlib==3.3.1
onnxruntime==1.22.0
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
orjson==3.10.18
packaging==25.0
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.10
proces==0.1.7
proglog==0.1.12
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
resampy==0.4.3
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.11.4
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
srt==3.5.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.14.1
torch==2.7.1
torchaudio==2.7.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
wasabi==1.1.3
weasel==0.4.1
websockets==11.0.3
Werkzeug==3.1.3
whisper-timestamped==1.14.2
wrapt==1.17.2
yarl==1.20.1
